Acceptable PR release
User is mentioned when code is posted in another channel.
Channel Link is provided when code is posted in another channel.
If use is in some sort of 'help' channel, code will be posted there, and
not moved to #programing_help.
Cleaned up the code to simplify it.
Added helper methods for logical code blocks.
Created array of code-like elements for the the bot to search. This can
easily be extended to include more code-like elements.
Other minor bug fixes.