Added functionality
CHANGELOG
* User is mentioned when code is posted in another channel.
* Channel Link is provided when code is posted in another channel.
* If user is in some sort of 'help' channel, code will be posted there,
and not moved to #programing_help.
* Cleaned up the code to simplify it.
* Algorithm should now catch all non-code lines and only format the
actual code. (may be cases that don't work, but have not found one yet.)
* Added helper methods for logical code blocks.
* Created array of code-like elements for the the bot to search. This
can easily be extended to include more code-like elements.
* Number of code elements is now a var, and can be edited to increase or
decrease the threshold for finding code.
* Better formatting on message.
* Bot will now remove the old, unformatted code from the channel (if it
has permission).
* If bot cannot delete message, an annoying message is posted directing
the user to ask that the bot be granted permission to delete messages.